ai-agent/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pyyaml>=6.0,<7.0
httpx>=0.27,<1.0
